+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.1] - 2025-11-27
+
+- fix: temporary publisher mapping for the ARD feed during the migration period
+- chore: update dependencies
+
 ## [2.1.0] - 2025-11-04
 
 - fix: prevent multiple downloads of the ard feed